Clarity to enable/disable delete.
When I first started into this, I wanted to make sure this was working, without deleteing anything.

I found it very helpful to let the lambda print out what it was going to do instead of actually deleting things.

Once I had a comfort level, it was chocks away and it all worked really nicely.